+++ websitebuild-20250429/alibabacloud_websitebuild20250429/models/_bind_app_domain_response_body.py
@@ -117,8 +117,6 @@ class BindAppDomainResponseBody(DaraModel):
 
         return self
 
-
-
 class BindAppDomainResponseBodyModule(DaraModel):
     def __init__(
         self,